refactor: move business logic and mockDb to services
