+++ src/controllers/auth.controller.ts
@@ -1,15 +1,7 @@
 // controllers/auth.controller.ts
-import bcrypt from 'bcrypt';
 import { Request, Response } from 'express';
-import jwt from 'jsonwebtoken';
 
-// Simulación de la base de datos
-const mockDb = {
-  User: {
-    findOne: async (query: any) => null, // Simula que no encuentra usuarios duplicados por defecto
-    create: async (data: any) => ({ id: 'new_user_123', ...data }),
-  },
-};
+import { AuthService } from '../services';
 
 export const register = async (req: Request, res: Response): Promise<void> => {
   try {
@@ -24,9 +16,10 @@ export const register = async (req: Request, res: Response): Promise<void> => {
     }
 
     // Verificar si el usuario ya existe en la base de datos
-    const existingUser = await mockDb.User.findOne({
-      $or: [{ email }, { username }],
-    });
+    const existingUser = await AuthService.findUserByEmailOrUsername(
+      email,
+      username,
+    );
     if (existingUser) {
       res
         .status(400)
@@ -34,16 +27,7 @@ export const register = async (req: Request, res: Response): Promise<void> => {
       return;
     }
 
-    // Hashear la contraseña con bcrypt (costo/salt rounds: 10) para cumplir con GDPR
-    const saltRounds = 10;
-    const hashedPassword = await bcrypt.hash(password, saltRounds);
-
-    // Crear y guardar el nuevo usuario
-    const newUser = await mockDb.User.create({
-      email,
-      username,
-      password: hashedPassword,
-    });
+    const newUser = await AuthService.registerUser(email, username, password);
 
     // Retornar respuesta exitosa (201 Created)
     res.status(201).json({
@@ -72,40 +56,19 @@ export const login = async (req: Request, res: Response): Promise<void> => {
       return;
     }
 
-    // Buscar al usuario en la base de datos
-    // Simulamos encontrar un usuario para que el flujo de login funcione en este ejemplo
-    const user = (await mockDb.User.findOne({ email })) || {
-      id: 'user_123',
-      username: 'PlayerOne',
-      email: email,
-      password: await bcrypt.hash('password123', 10), // Contraseña hasheada simulada
-    };
+    // Intentar loguear al usuario usando el servicio
+    const authData = await AuthService.loginUser(email, password);
 
-    if (!user) {
+    if (!authData) {
       res.status(401).json({ message: 'Credenciales inválidas.' });
       return;
     }
-
-    // Comparar la contraseña ingresada con el hash de la base de datos
-    const isPasswordValid = await bcrypt.compare(password, user.password);
-    if (!isPasswordValid) {
-      res.status(401).json({ message: 'Credenciales inválidas.' });
-      return;
-    }
-
-    // Generar el token JWT
-    const secretKey = process.env.JWT_SECRET || 'super_secret_fallback_key';
-    const token = jwt.sign(
-      { id: user.id, username: user.username },
-      secretKey,
-      { expiresIn: '24h' }, // El token expira en 24 horas
-    );
 
     // Retornar el token y los datos del usuario
     res.status(200).json({
       message: 'Inicio de sesión exitoso.',
-      token,
-      user: { id: user.id, username: user.username },
+      token: authData.token,
+      user: { id: authData.user.id, username: authData.user.username },
     });
   } catch (error) {
     console.error('Error in login:', error);
