+++ src/controllers/user.controller.ts
@@ -1,54 +1,8 @@
 // controllers/user.controller.ts
 import { Response } from 'express';
 
+import { UserService } from '../services';
 import { AuthenticatedRequest } from '../types';
-
-// Simulación de las llamadas a la base de datos
-export const mockDb = {
-  User: {
-    findById: async (id: string) => ({
-      id,
-      username: 'PlayerOne',
-      level: 15,
-      status: 'online',
-    }),
-    find: async (query: any) => [{ id: 'user_999', username: 'TestUser' }],
-  },
-  Economy: {
-    findByUserId: async (id: string) => ({ coins: 2500, gems: 150 }),
-  },
-  Inventory: {
-    findByUserId: async (id: string) => [
-      { itemId: 'p1', name: 'Rastreador', quantity: 2 },
-    ],
-  },
-  CardCollection: {
-    findOwnedByUserId: async (id: string) => [
-      { cardId: 'c_001', name: 'Golpe Crítico', quantity: 3 },
-    ],
-  },
-  Deck: {
-    find: async (query: any) => [
-      {
-        deckId: 'd_123',
-        name: 'Mazo Destrucción',
-        cardCount: 40,
-        userId: 'user_123',
-      },
-    ],
-    create: async (data: any) => ({ deckId: `d_${Date.now()}`, ...data }),
-    findById: async (id: string) => ({
-      deckId: id,
-      userId: 'user_123',
-      name: 'Mazo Destrucción',
-    }), // Simula que el dueño es user_123
-    findByIdAndUpdate: async (id: string, data: any) => ({
-      deckId: id,
-      ...data,
-    }),
-    findByIdAndDelete: async (id: string) => true,
-  },
-};
 
 export const getProfile = async (
   req: AuthenticatedRequest,
@@ -56,7 +10,7 @@ export const getProfile = async (
 ): Promise<void> => {
   try {
     const userId = req.user!.id; // El middleware garantiza que req.user existe
-    const userProfile = await mockDb.User.findById(userId);
+    const userProfile = await UserService.getUserProfile(userId);
 
     if (!userProfile) {
       res.status(404).json({ message: 'Perfil de usuario no encontrado.' });
@@ -77,7 +31,7 @@ export const getBalance = async (
 ): Promise<void> => {
   try {
     const userId = req.user!.id;
-    const balance = await mockDb.Economy.findByUserId(userId);
+    const balance = await UserService.getUserEconomy(userId);
 
     res.status(200).json({ balance });
   } catch (error) {
@@ -91,7 +45,7 @@ export const getInventory = async (
 ): Promise<void> => {
   try {
     const userId = req.user!.id;
-    const inventory = await mockDb.Inventory.findByUserId(userId);
+    const inventory = await UserService.getUserInventory(userId);
 
     res.status(200).json({ inventory });
   } catch (error) {
@@ -116,9 +70,7 @@ export const searchUsers = async (
     }
 
     // Búsqueda en la base de datos (simulada)
-    const results = await mockDb.User.find({
-      username: { $regex: searchQuery, $options: 'i' },
-    });
+    const results = await UserService.searchUsers(searchQuery);
 
     res.status(200).json({ results });
   } catch (error) {
@@ -132,7 +84,7 @@ export const getOwnedCards = async (
 ): Promise<void> => {
   try {
     const userId = req.user!.id;
-    const ownedCards = await mockDb.CardCollection.findOwnedByUserId(userId);
+    const ownedCards = await UserService.getUserCards(userId);
 
     res.status(200).json({ cards: ownedCards });
   } catch (error) {
@@ -148,7 +100,7 @@ export const getUserDecks = async (
 ): Promise<void> => {
   try {
     const userId = req.user!.id;
-    const decks = await mockDb.Deck.find({ userId });
+    const decks = await UserService.getUserDecks(userId);
 
     res.status(200).json({ decks });
   } catch (error) {
@@ -166,13 +118,7 @@ export const createDeck = async (
     const userId = req.user!.id;
     const { name, cardIds } = req.body;
 
-    const newDeck = await mockDb.Deck.create({
-      userId,
-      name: name.trim(),
-      cards: cardIds,
-      cardCount: cardIds.length,
-    });
-
+    const newDeck = await UserService.createDeck(userId, name, cardIds);
     res
       .status(201)
       .json({ message: 'Mazo creado exitosamente.', deck: newDeck });
@@ -189,11 +135,7 @@ export const updateDeck = async (
     const { deckId } = req.params;
     const { name, cardIds } = req.body;
 
-    const updatedDeck = await mockDb.Deck.findByIdAndUpdate(deckId, {
-      name: name.trim(),
-      cards: cardIds,
-    });
-
+    const updatedDeck = await UserService.updateDeck(deckId, name, cardIds);
     res.status(200).json({ message: 'Mazo actualizado.', deck: updatedDeck });
   } catch (error) {
     res.status(500).json({ message: 'Error interno al actualizar.' });
@@ -208,8 +150,7 @@ export const deleteDeck = async (
     const { deckId } = req.params;
 
     // El middleware garantizó que es su mazo
-    await mockDb.Deck.findByIdAndDelete(deckId);
-
+    await UserService.deleteDeck(deckId);
     res.status(200).json({ message: 'Mazo eliminado correctamente.' });
   } catch (error) {
     res.status(500).json({ message: 'Error al eliminar el mazo.' });
