+++ src/services/index.ts
@@ -1,0 +1,5 @@
+export * from './auth.service';
+export * from './collection.service';
+export * from './friend.service';
+export * from './shop.service';
+export * from './user.service';
